refactor: split YieldAnalysis.tsx into DscrBadge/YieldGauge/CustomScenarioSlider/StressScenarioTable
